create universal types for breeds, favourites, images types
