+++ src/types/common.tsx
@@ -1,50 +1,11 @@
 export type QueryParams = Record<string, string>;
 
-export type Breed = {
-  weight: {
-    imperial: string;
-    metric: string;
-  };
+export type Weight = {
+  imperial: string;
+  metric: string;
+};
+
+export type Categorie = {
   id: string;
   name: string;
-  cfa_url: string;
-  vetstreet_url: string;
-  vcahospitals_url: string;
-  temperament: string;
-  origin: string;
-  country_codes: string;
-  country_code: string;
-  description: string;
-  life_span: string;
-  indoor: string;
-  lap: string;
-  alt_names: string;
-  adaptability: string;
-  affection_level: string;
-  child_friendly: string;
-  dog_friendly: string;
-  energy_level: string;
-  grooming: string;
-  health_issues: string;
-  intelligence: string;
-  shedding_level: string;
-  social_needs: string;
-  stranger_friendly: string;
-  vocalisation: string;
-  experimental: string;
-  hairless: string;
-  natural: string;
-  rare: string;
-  rex: string;
-  suppressed_tail: string;
-  short_legs: string;
-  wikipedia_url: string;
-  hypoallergenic: string;
-  reference_image_id: string;
-  image: {
-    id: string;
-    width: string;
-    height: string;
-    url: string;
-  };
 };
